refactor complete (for now ;)
