+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 import { ApolloProvider } from '@apollo/client';
 import { client } from './utils/apolloClient';
-import { RepoPage } from './components/organisms/repoTable';
+import { RepoPage } from './components/organisms/repoPage';
 
 const App = (): JSX.Element => {
   return (
