+++ src/components/organisms/repoPage.tsx
@@ -1,36 +1,8 @@
 import { useLazyQuery } from '@apollo/client';
 import { useState, useEffect } from 'react';
 import { REPO_SEARCH } from '../../queries/queries';
-import { QueryResult, EdgesEntity } from '../../types/repoQuery';
-import { RepoRow } from '../atoms/repoRow';
-
-export interface RepoTableProps {
-  resultSet: EdgesEntity[] | null | undefined;
-}
-
-export const RepoTable = ({ resultSet }: RepoTableProps): JSX.Element => {
-  console.log(resultSet);
-  if (resultSet === null || resultSet === undefined) {
-    return <h1>no results</h1>;
-  }
-  return (
-    <div>
-      {resultSet.map((res, idx) => {
-        return (
-          <div key={res.node.id}>
-            <RepoRow
-              idx={idx}
-              title={res.node.name}
-              url={res.node.url}
-              stars={`${res.node.stargazerCount}`}
-              forks={`${res.node.forkCount}`}
-            />
-          </div>
-        );
-      })}
-    </div>
-  );
-};
+import { QueryResult } from '../../types/repoQuery';
+import { RepoTable } from '../molecules/repoTable';
 
 export const RepoPage = (): JSX.Element => {
   const repoQuery = 'react';
